feat: support more citation key characters
This commit expands the regex pattern used to validate citation keys,
thus reducing the number of citation keys that mkdoc-bibtex would
previously reject, even though they were valid (according to
pandoc and BibTex).

Here's how the correct syntax for valid citation keys was derived:

1. According to these StackExchange posts:

https://tex.stackexchange.com/questions/408530/what-characters-are-allowed-to-use-as-delimiters-for-bibtex-keys
https://tex.stackexchange.com/questions/96454/using-bibtex-keys-containing-parentheses-with-biber/96918#96918

the following punctuation symbols are DISALLOWED:

"#'(),={}%~\

2. Pandoc specifies the following rules:

https://pandoc.org/MANUAL.html#citation-syntax

Which imply that punctuation except for underscore can not be repeated
in keys. It also defines the following set of ALLOWED punctuation
characters:

:.#$%&-+?<>~/

However, experimentatin with pandoc suggests that the following
characters are always DISALLOWED:

%#<>~ ^!|

Admittedly, the final three were not on the original list of allowed characters.

Finally, pandoc also requires that the key start with an alphanumeric
symbol or underscore, and that the allowed punctuation characters do NOT
appear as the first or final symbol in a key, and that the allowed
punctuation characters only appear in groups of one. This leads to the
following regex for citation keys which should be accepted by all
systems used by mkdocs-bibtex:

r"@(?P<key>[\w0-9_]([:.$&+?/-]?[\w0-9_]+)*)"